重构main.py拆分函数到其它几个模块中
