+++ main.py
@@ -14,42 +14,16 @@ import time
 import zipfile
 from concurrent.futures import ThreadPoolExecutor, as_completed
 from datetime import datetime
-from urllib.parse import urlparse, parse_qs, unquote
 
 import requests
-import yaml
 
 from common import constants
+from common.constants import TEST_URLS, DEBUG_MODE, CONNECTION_TIMEOUT
 from common.decorators import timer
 from common.logger import Logger
 from config.settings import Settings
+from subscription import Subscriber
 from utils import utils
-
-# 支持的协议类型列表
-SUPPORTED_PROTOCOLS = [
-    'vmess://',
-    'trojan://',
-    'vless://',
-    'ss://',
-    'ssr://',
-    'http://',
-    'https://',
-    'socks://',
-    'socks5://',
-    'hysteria://',
-    'wireguard://'
-]
-
-# 测速相关配置
-# 测试URL列表
-TEST_URLS = [
-    "http://www.gstatic.com/generate_204",  # Google测试
-]
-CONNECTION_TIMEOUT = 10  # 连接超时时间，单位为秒
-DEBUG_MODE = False  # 默认开启调试模式，方便查看处理过程
-
-# 核心程序配置
-CORE_PATH = None  # 核心程序路径，将自动检测
 
 
 def is_github_raw_url(url):
@@ -393,590 +367,6 @@ def fetch_content(url):
         import traceback
         traceback.print_exc()
         return None
-
-
-def parse_clash_yaml(content):
-    """解析Clash配置文件"""
-    try:
-        data = yaml.safe_load(content)
-        if not data:
-            return []
-
-        # 直接查找proxies字段，无论它在哪个层级
-        if 'proxies' in data:
-            if DEBUG_MODE:
-                logging.info(f"从YAML中找到 {len(data['proxies'])} 个节点")
-            return data['proxies']
-
-        # 如果没有找到proxies字段，尝试其他可能的字段名
-        for key in ['proxy-providers', 'Proxy', 'proxys']:
-            if key in data and isinstance(data[key], list):
-                if DEBUG_MODE:
-                    logging.info(f"从YAML的{key}字段中找到 {len(data[key])} 个节点")
-                return data[key]
-
-        logging.info("YAML中未找到节点信息")
-        return []
-    except Exception as e:
-        # logging.info(f"解析Clash YAML失败: {str(e)}")
-        return []
-
-
-def parse_v2ray_base64(content):
-    """解析V2Ray Base64编码的配置"""
-    try:
-        # 处理多行base64
-        content = content.strip().replace('\n', '').replace('\r', '')
-        # 尝试修复可能的编码问题
-        try:
-            if isinstance(content, bytes):
-                content = content.decode('utf-8')
-            # 确保内容是ASCII兼容的
-            content = content.encode('ascii', 'ignore').decode('ascii')
-        except UnicodeError:
-            logging.info("Error: Invalid encoding in base64 content")
-            return []
-
-        try:
-            decoded = base64.b64decode(content + '=' * (-len(content) % 4))
-            decoded_str = decoded.decode('utf-8', 'ignore')
-        except Exception as e:
-            logging.info(f"Error decoding base64 content: {str(e)}")
-            return []
-
-        nodes = []
-        for line in decoded_str.split('\n'):
-            if line.startswith('vmess://') or line.startswith('trojan://'):
-                node = parse_v2ray_uri(line)
-                if node:
-                    nodes.append(node)
-        return nodes
-    except Exception as e:
-        # logging.info(f"Error parsing V2Ray base64: {str(e)}")
-        return []
-
-
-def parse_v2ray_uri(uri):
-    """解析V2Ray URI格式的配置"""
-    try:
-        # 处理vmess协议
-        if uri.startswith('vmess://'):
-            b64_config = uri.replace('vmess://', '')
-            # 确保base64正确填充
-            b64_config = b64_config + '=' * (-len(b64_config) % 4)
-            try:
-                config = json.loads(base64.b64decode(b64_config).decode())
-                return {
-                    'type': 'vmess',
-                    'name': config.get('ps', 'Unknown'),
-                    'server': config.get('add', ''),
-                    'port': int(config.get('port', 0)),
-                    'uuid': config.get('id', ''),
-                    'alterId': int(config.get('aid', 0)),
-                    'cipher': config.get('type', 'auto'),
-                    'tls': config.get('tls', '') == 'tls',
-                    'network': config.get('net', 'tcp')
-                }
-            except json.JSONDecodeError:
-                # 某些情况下vmess可能使用非标准格式
-                logging.info(f"Non-standard vmess format: {uri}")
-                return None
-
-        # 处理trojan协议
-        elif uri.startswith('trojan://'):
-            parsed = urlparse(uri)
-            query = parse_qs(parsed.query)
-            return {
-                'type': 'trojan',
-                'name': query.get('sni', [query.get('peer', ['Unknown'])[0]])[0],
-                'server': parsed.hostname or '',
-                'port': parsed.port or 443,
-                'password': parsed.username or ''
-            }
-
-        # 处理vless协议
-        elif uri.startswith('vless://'):
-            parsed = urlparse(uri)
-            query = parse_qs(parsed.query)
-            return {
-                'type': 'vless',
-                'name': query.get('remarks', [query.get('sni', ['Unknown'])[0]])[0],
-                'server': parsed.hostname or '',
-                'port': parsed.port or 443,
-                'uuid': parsed.username or '',
-                'tls': query.get('security', [''])[0] == 'tls',
-                'flow': query.get('flow', [''])[0],
-                'network': query.get('type', ['tcp'])[0]
-            }
-
-        # 处理shadowsocks协议
-        elif uri.startswith('ss://'):
-            # 首先获取#后面的名称部分（如果存在）
-            name = 'Unknown'
-            if '#' in uri:
-                name_part = uri.split('#', 1)[1]
-                name = unquote(name_part)
-                uri = uri.split('#', 1)[0]  # 移除名称部分以便后续处理
-
-            if '@' in uri:
-                # 处理 ss://method:password@host:port
-                parsed = urlparse(uri)
-                server = parsed.hostname
-                port = parsed.port
-
-                # 提取方法和密码
-                userinfo = parsed.username
-                if userinfo:
-                    try:
-                        # 有些实现可能会对userinfo进行base64编码
-                        decoded = base64.b64decode(userinfo + '=' * (-len(userinfo) % 4)).decode()
-                        if ':' in decoded:
-                            method, password = decoded.split(':', 1)
-                        else:
-                            method, password = 'aes-256-gcm', userinfo
-                    except:
-                        # 如果不是base64编码，可能是明文
-                        if ':' in userinfo:
-                            method, password = userinfo.split(':', 1)
-                        else:
-                            method, password = 'aes-256-gcm', userinfo
-                else:
-                    method, password = 'aes-256-gcm', ''
-
-                # 如果查询参数中包含remarks，优先使用它
-                query = parse_qs(parsed.query)
-                if 'remarks' in query:
-                    name = query.get('remarks', ['Unknown'])[0]
-
-                return {
-                    'type': 'ss',
-                    'name': name,
-                    'server': server or '',
-                    'port': port or 443,
-                    'cipher': method,
-                    'password': password
-                }
-            else:
-                # 处理 ss://BASE64(method:password@host:port)
-                b64_config = uri.replace('ss://', '')
-                try:
-                    # 确保base64正确填充
-                    b64_config = b64_config + '=' * (-len(b64_config) % 4)
-
-                    config_str = base64.b64decode(b64_config).decode()
-                    # 提取方法和密码
-                    if '@' in config_str:
-                        method_pwd, server_port = config_str.rsplit('@', 1)
-                        method, password = method_pwd.split(':', 1)
-                        server, port = server_port.rsplit(':', 1)
-
-                        return {
-                            'type': 'ss',
-                            'name': name,
-                            'server': server,
-                            'port': int(port),
-                            'cipher': method,
-                            'password': password
-                        }
-                except Exception as e:
-                    # logging.info(f"Invalid ss URI format: {uri}, error: {str(e)}")
-                    return None
-
-        # 处理shadowsocksr协议
-        elif uri.startswith('ssr://'):
-            b64_config = uri.replace('ssr://', '')
-            try:
-                # 确保base64正确填充
-                b64_config = b64_config + '=' * (-len(b64_config) % 4)
-                config_str = base64.b64decode(b64_config).decode()
-
-                # SSR格式: server:port:protocol:method:obfs:base64pass/?obfsparam=base64param&protoparam=base64param
-                # &remarks=base64remarks
-                parts = config_str.split(':')
-                if len(parts) >= 6:
-                    server = parts[0]
-                    port = parts[1]
-                    protocol = parts[2]
-                    method = parts[3]
-                    obfs = parts[4]
-
-                    # 处理剩余参数
-                    password_and_params = parts[5].split('/?', 1)
-                    password_b64 = password_and_params[0]
-                    password = base64.b64decode(password_b64 + '=' * (-len(password_b64) % 4)).decode()
-
-                    # 提取参数
-                    name = 'Unknown'
-                    if len(password_and_params) > 1 and 'remarks=' in password_and_params[1]:
-                        remarks_b64 = password_and_params[1].split('remarks=', 1)[1].split('&', 1)[0]
-                        try:
-                            name = base64.b64decode(remarks_b64 + '=' * (-len(remarks_b64) % 4)).decode()
-                        except:
-                            pass
-
-                    return {
-                        'type': 'ssr',
-                        'name': name,
-                        'server': server,
-                        'port': int(port),
-                        'protocol': protocol,
-                        'cipher': method,
-                        'obfs': obfs,
-                        'password': password
-                    }
-            except Exception as e:
-                # logging.info(f"Error parsing SSR URI: {str(e)}")
-                return None
-
-        # 处理HTTP/HTTPS协议
-        elif uri.startswith(('http://', 'https://')):
-            parsed = urlparse(uri)
-            query = parse_qs(parsed.query)
-            return {
-                'type': 'http' if uri.startswith('http://') else 'https',
-                'name': query.get('remarks', ['Unknown'])[0],
-                'server': parsed.hostname or '',
-                'port': parsed.port or (80 if uri.startswith('http://') else 443),
-                'username': parsed.username or '',
-                'password': parsed.password or ''
-            }
-
-        # 处理SOCKS协议
-        elif uri.startswith(('socks://', 'socks5://')):
-            parsed = urlparse(uri)
-            query = parse_qs(parsed.query)
-            return {
-                'type': 'socks',
-                'name': query.get('remarks', ['Unknown'])[0],
-                'server': parsed.hostname or '',
-                'port': parsed.port or 1080,
-                'username': parsed.username or '',
-                'password': parsed.password or ''
-            }
-
-        # 处理hysteria协议
-        elif uri.startswith('hysteria://'):
-            parsed = urlparse(uri)
-            query = parse_qs(parsed.query)
-            return {
-                'type': 'hysteria',
-                'name': query.get('peer', ['Unknown'])[0],
-                'server': parsed.hostname or '',
-                'port': parsed.port or 443,
-                'protocol': query.get('protocol', [''])[0],
-                'auth': parsed.username or query.get('auth', [''])[0]
-            }
-
-        # 处理wireguard协议
-        elif uri.startswith('wireguard://'):
-            parsed = urlparse(uri)
-            query = parse_qs(parsed.query)
-            return {
-                'type': 'wireguard',
-                'name': query.get('remarks', ['Unknown'])[0],
-                'server': parsed.hostname or '',
-                'port': parsed.port or 51820,
-                'private_key': query.get('privateKey', [''])[0],
-                'public_key': query.get('publicKey', [''])[0],
-                'allowed_ips': query.get('allowedIPs', ['0.0.0.0/0'])[0]
-            }
-
-    except Exception as e:
-        # logging.info(f"Error parsing URI: {str(e)}")
-        return None
-
-
-def extract_nodes(content):
-    """级联提取节点，按照Base64 -> YAML -> 正则表达式 -> JSON的顺序尝试"""
-    if not content:
-        return []
-
-    nodes = []
-    methods_tried = []
-
-    # 1. 尝试Base64解码提取
-    try:
-        # 处理多行base64，移除所有空白字符和特殊字符
-        cleaned_content = re.sub(r'[\s\n\r\t]+', '', content)
-        cleaned_content = re.sub(r'[^A-Za-z0-9+/=]', '', cleaned_content)
-
-        # 确保base64字符串长度是4的倍数
-        padding_length = len(cleaned_content) % 4
-        if padding_length:
-            cleaned_content += '=' * (4 - padding_length)
-
-        # 尝试base64解码
-        try:
-            decoded_bytes = base64.b64decode(cleaned_content)
-            decoded_str = decoded_bytes.decode('utf-8', 'ignore')
-
-            # 检查解码后的内容是否包含任何支持的协议节点
-            if any(protocol in decoded_str for protocol in SUPPORTED_PROTOCOLS):
-                logging.info("使用Base64解码提取节点")
-                methods_tried.append("Base64")
-                for line in decoded_str.split('\n'):
-                    line = line.strip()
-                    if any(line.startswith(protocol) for protocol in SUPPORTED_PROTOCOLS):
-                        node = parse_v2ray_uri(line)
-                        if node:
-                            nodes.append(node)
-        except Exception as e:
-            # logging.info(f"Base64解码失败或未找到节点: {str(e)}")
-            pass
-    except Exception as e:
-        logging.info(f"Base64预处理失败: {str(e)}")
-
-    # 如果已经提取到节点，直接返回
-    if len(nodes) > 0:
-        logging.info(f"通过【{methods_tried[-1]}】方法成功提取到{len(nodes)}个节点")
-        return nodes
-
-    # 2. 尝试解析YAML格式
-    try:
-        # 移除HTML标签和特殊标记
-        cleaned_content = re.sub(r'<[^>]+>|!&lt;str&gt;', '', content)
-
-        # 更强大的YAML格式检测，查找常见Clash配置特征
-        yaml_indicators = [
-            'proxies:', 'Proxy:', 'proxy:', 'proxy-providers:',
-            'port:', 'socks-port:', 'allow-lan:', 'mode:',
-            'type: vmess', 'type: ss', 'type: trojan', 'type: vless'
-        ]
-
-        if any(indicator in cleaned_content for indicator in yaml_indicators):
-            # logging.info("尝试解析YAML格式内容")
-            methods_tried.append("YAML")
-
-            # 尝试直接加载YAML
-            try:
-                yaml_nodes = parse_clash_yaml(cleaned_content)
-                if yaml_nodes:
-                    # logging.info(f"从YAML中提取到{len(yaml_nodes)}个节点")
-                    nodes.extend(yaml_nodes)
-            except Exception as yaml_error:
-                logging.info(f"标准YAML解析失败: {str(yaml_error)}")
-
-                # 如果标准解析失败，尝试更宽松的解析方式
-                try:
-                    # 尝试提取proxies部分
-                    proxies_match = re.search(r'proxies:\s*\n([\s\S]+?)(?:\n\w+:|$)', cleaned_content)
-                    if proxies_match:
-                        proxies_yaml = "proxies:\n" + proxies_match.group(1)
-                        yaml_nodes = parse_clash_yaml(proxies_yaml)
-                        if yaml_nodes:
-                            logging.info(f"从proxies块提取到{len(yaml_nodes)}个节点")
-                            nodes.extend(yaml_nodes)
-                except Exception as fallback_error:
-                    logging.info(f"尝试解析proxies块失败: {str(fallback_error)}")
-    except Exception as e:
-        logging.info(f"YAML解析过程出错: {str(e)}")
-
-    # 如果已经提取到节点，直接返回
-    if len(nodes) > 0:
-        logging.info(f"通过【{methods_tried[-1]}】方法成功提取到{len(nodes)}个节点")
-        return nodes
-
-    # 3. 尝试使用正则表达式直接提取
-    try:
-        # logging.info("尝试使用正则表达式直接提取节点")
-        methods_tried.append("正则表达式")
-
-        # 为每种支持的协议定义正则表达式并提取
-        for protocol in SUPPORTED_PROTOCOLS:
-            if protocol == 'vmess://':
-                # vmess通常是一个base64编码的字符串
-                found_nodes = re.findall(r'vmess://[A-Za-z0-9+/=]+', content)
-            elif protocol == 'hysteria://' or protocol == 'wireguard://':
-                # 这些协议可能有特殊格式，需要特别处理
-                found_nodes = re.findall(f'{protocol}[^"\'<>\\s]+', content)
-            else:
-                # 对于其他协议，采用通用正则表达式
-                found_nodes = re.findall(f'{protocol}[^"\'<>\\s]+', content)
-
-            for uri in found_nodes:
-                node = parse_v2ray_uri(uri)
-                if node:
-                    nodes.append(node)
-    except Exception as e:
-        logging.info(f"正则表达式提取失败: {str(e)}")
-
-    # 如果已经提取到节点，直接返回
-    if len(nodes) > 0:
-        logging.info(f"通过【{methods_tried[-1]}】方法成功提取到{len(nodes)}个节点")
-        return nodes
-
-    # 4. 尝试解析JSON格式
-    try:
-        # logging.info("尝试解析JSON格式")
-        methods_tried.append("JSON")
-
-        # 清理内容，移除可能的HTML标签和注释
-        cleaned_content = re.sub(r'<[^>]+>|/\*.*?\*/|//.*?$', '', content, flags=re.MULTILINE)
-
-        # 尝试解析JSON
-        try:
-            json_data = json.loads(cleaned_content)
-            json_nodes = parse_json_nodes(json_data)
-            if json_nodes:
-                # logging.info(f"从JSON中提取到{len(json_nodes)}个节点")
-                nodes.extend(json_nodes)
-        except json.JSONDecodeError as e:
-            # 尝试查找内容中的JSON片段
-            try:
-                # 查找类似于 [{...}] 或 {...} 形式的JSON
-                json_matches = re.findall(r'(\[{.*?}]|\{.*?})', cleaned_content, re.DOTALL)
-                for json_match in json_matches:
-                    try:
-                        potential_json = json.loads(json_match)
-                        json_nodes = parse_json_nodes(potential_json)
-                        if json_nodes:
-                            # logging.info(f"从JSON片段中提取到{len(json_nodes)}个节点")
-                            nodes.extend(json_nodes)
-                            # 找到有效的JSON片段后，不再继续查找
-                            break
-                    except:
-                        continue
-            except Exception as extract_error:
-                # logging.info(f"尝试提取JSON片段失败: {str(extract_error)}")
-                pass
-    except Exception as e:
-        logging.info(f"JSON解析过程出错: {str(e)}")
-
-    if len(nodes) > 0:
-        logging.info(f"通过【{methods_tried[-1]}】方法成功提取到{len(nodes)}个节点")
-        return nodes
-    else:
-        logging.info("未找到任何节点")
-        return []
-
-
-def parse_json_nodes(json_data):
-    """从JSON数据中解析节点信息"""
-    nodes = []
-
-    # 处理数组形式的JSON
-    if isinstance(json_data, list):
-        for item in json_data:
-            node = parse_single_json_node(item)
-            if node:
-                nodes.append(node)
-    # 处理对象形式的JSON
-    elif isinstance(json_data, dict):
-        # 检查是否是单个节点
-        node = parse_single_json_node(json_data)
-        if node:
-            nodes.append(node)
-        # 检查是否包含节点列表
-        elif 'servers' in json_data and isinstance(json_data['servers'], list):
-            for server in json_data['servers']:
-                node = parse_single_json_node(server)
-                if node:
-                    nodes.append(node)
-        # 检查其他可能的字段名
-        for key in ['proxies', 'nodes', 'configs']:
-            if key in json_data and isinstance(json_data[key], list):
-                for item in json_data[key]:
-                    node = parse_single_json_node(item)
-                    if node:
-                        nodes.append(node)
-
-    return nodes
-
-
-def parse_single_json_node(item):
-    """解析单个JSON节点数据"""
-    # 如果不是字典，直接返回
-    if not isinstance(item, dict):
-        return None
-
-    # 支持Shadowsocks格式
-    if ('server' in item and 'server_port' in item and
-            'method' in item and 'password' in item):
-        try:
-            return {
-                'type': 'ss',
-                'name': item.get('remarks', f"SS-{item['server']}"),
-                'server': item['server'],
-                'port': int(item['server_port']),
-                'cipher': item['method'],
-                'password': item['password'],
-                'plugin': item.get('plugin', ''),
-                'plugin_opts': item.get('plugin_opts', '')
-            }
-        except Exception as e:
-            logging.info(f"解析Shadowsocks节点失败: {str(e)}")
-            return None
-
-    # 支持VMess格式
-    elif 'add' in item and 'port' in item and 'id' in item:
-        try:
-            return {
-                'type': 'vmess',
-                'name': item.get('ps', item.get('remarks', f"VMess-{item['add']}")),
-                'server': item['add'],
-                'port': int(item['port']),
-                'uuid': item['id'],
-                'alterId': int(item.get('aid', 0)),
-                'cipher': item.get('scy', item.get('security', 'auto')),
-                'tls': item.get('tls', '') == 'tls',
-                'network': item.get('net', 'tcp'),
-                'path': item.get('path', '/'),
-                'host': item.get('host', '')
-            }
-        except Exception as e:
-            logging.info(f"解析VMess节点失败: {str(e)}")
-            return None
-
-    # 支持Trojan格式
-    elif ('server' in item and 'port' in item and 'password' in item and
-          item.get('type', '').lower() == 'trojan'):
-        try:
-            return {
-                'type': 'trojan',
-                'name': item.get('remarks', f"Trojan-{item['server']}"),
-                'server': item['server'],
-                'port': int(item['port']),
-                'password': item['password'],
-                'sni': item.get('sni', item.get('peer', ''))
-            }
-        except Exception as e:
-            logging.info(f"解析Trojan节点失败: {str(e)}")
-            return None
-
-    # 支持Clash格式
-    elif 'type' in item and 'server' in item and 'port' in item:
-        try:
-            node_type = item['type'].lower()
-            if node_type in ['ss', 'vmess', 'trojan', 'vless', 'http', 'socks']:
-                node = {
-                    'type': node_type,
-                    'name': item.get('name', f"{node_type.upper()}-{item['server']}"),
-                    'server': item['server'],
-                    'port': int(item['port'])
-                }
-
-                # 根据不同类型添加特定字段
-                if node_type == 'ss':
-                    node['cipher'] = item.get('cipher', 'aes-256-gcm')
-                    node['password'] = item.get('password', '')
-                elif node_type == 'vmess':
-                    node['uuid'] = item.get('uuid', '')
-                    node['alterId'] = int(item.get('alterId', 0))
-                    node['cipher'] = item.get('cipher', 'auto')
-                    node['tls'] = item.get('tls', False)
-                    node['network'] = item.get('network', 'tcp')
-                    if 'ws-path' in item:
-                        node['path'] = item['ws-path']
-                elif node_type in ['trojan', 'vless']:
-                    node['password'] = item.get('password', '')
-                    node['sni'] = item.get('sni', '')
-
-                return node
-        except Exception as e:
-            logging.info(f"解析Clash节点失败: {str(e)}")
-            return None
-
-    return None
 
 
 def download_xray_core():
@@ -1530,65 +920,6 @@ def process_node(node):
     return node
 
 
-
-
-
-def load_subscriptions(config: dict) -> list[str]:
-    """加载并合并配置中的订阅链接和汇聚订阅链接，去重后返回列表。"""
-    subs = config.get("subscriptions", [])
-    agg_url = config.get("aggSubs")
-
-    if agg_url:
-        resp = requests.get(agg_url, timeout=10)
-        if resp.ok:
-            subs.extend(resp.text.splitlines())
-
-    # 去重且保持顺序
-    return list(dict.fromkeys(subs))
-
-
-def _fetch_and_extract(link: str) -> list[dict]:
-    """
-    线程中运行：拉取订阅、提取节点列表。
-    返回一个节点字典列表，失败时返回空列表。
-    """
-    try:
-        with requests.Session() as session:
-            resp = session.get(link, timeout=10)
-            resp.raise_for_status()
-            nodes = extract_nodes(resp.text)
-            logging.info(f"  ✓ 完成：{link}，提取 {len(nodes)} 个节点")
-            return nodes
-    except Exception as e:
-        logging.warning(f"[WARN] 拉取/解析失败：{link} -> {e}")
-        return []
-
-
-@timer(unit="ms")
-def gather_all_nodes(sub_links: list[str], max_workers: int | None = None) -> list[dict]:
-    """
-    并发拉取并解析所有订阅链接（I/O 密集型），返回聚合后的节点列表。
-    """
-    max_workers = Settings.THREAD_POOL_SIZE if max_workers is None else max_workers
-    logging.info(f"开始并发获取节点信息，使用线程池最大并发数：{max_workers}")
-    all_nodes: list[dict] = []
-
-    # 不指定 max_workers 时，ThreadPoolExecutor 会使用 min(32, os.cpu_count() + 4)
-    with ThreadPoolExecutor(max_workers=max_workers) as executor:
-        future_to_link = {executor.submit(_fetch_and_extract, link): link for link in sub_links}
-        for future in as_completed(future_to_link):
-            # 即使 _fetch_and_extract 已经捕获了异常，这里也做一层保险
-            try:
-                nodes = future.result()
-                all_nodes.extend(nodes)
-            except Exception as e:
-                link = future_to_link[future]
-                logging.error(f"[ERROR] 处理 {link} 的线程异常：{e}")
-
-    logging.info(f"全部完成，总计提取节点：{len(all_nodes)}")
-    return all_nodes
-
-
 @timer(unit="ms")
 def _test_all_nodes_latency(
         nodes: list[dict],
@@ -1651,13 +982,10 @@ def main():
     config = Settings().config
     core_path = find_core_program()
     logging.info(f"核心程序路径：{core_path}")
-
-    # 1. 加载并去重订阅
-    sub_links = load_subscriptions(config)
-    logging.info(f"共 {len(sub_links)} 条订阅链接")
+    subscriber = Subscriber(config)
 
     # 2. 获取并解析所有节点
-    all_nodes = gather_all_nodes(sub_links)
+    all_nodes = subscriber.get_subscription_nodes()
     logging.info(f"提取到节点总数：{len(all_nodes)}")
 
     # 3. 去重
